backend/requirements.txt
Django>=4.2
djangorestframework
pytest
pytest-django
pytz
